+++ src/parser/niseci/mod.rs
@@ -620,8 +620,11 @@ pub fn parse_records_anagrafica_niseci<T: RecordAnagraficaNISECI>(
 
     let res = AnagraficaNISECI::new(
         ComunitaNISECI {
+            #[expect(deprecated)]
             tipo: tipo_comunita,
+            #[expect(deprecated)]
             fonte: Some(r.fonte()),
+            #[expect(deprecated)]
             numero_protocollo: Some(r.numero_protocollo()),
         },
         r.codice_stazione(),