+++ src/tests/engines/niseci/x2/mod.rs
@@ -106,8 +106,11 @@ fn calculate_x2_test_1() {
     let campionamento = create_massive_campionamento_ciacci();
 
     let comunita = ComunitaNISECI {
+        #[expect(deprecated)]
         fonte: Some("hey".to_string()),
+        #[expect(deprecated)]
         numero_protocollo: None,
+        #[expect(deprecated)]
         tipo: TipoComunitaNISECI::Dm260_2010,
     };
 
@@ -139,8 +142,11 @@ fn calculate_x2_test_1() {
     let campionamento = create_massive_campionamento_ciacci_2();
 
     let comunita = ComunitaNISECI {
+        #[expect(deprecated)]
         fonte: Some("hey".to_string()),
+        #[expect(deprecated)]
         numero_protocollo: None,
+        #[expect(deprecated)]
         tipo: TipoComunitaNISECI::Dm260_2010,
     };